SQUASH: consistently use size_t len and uint64_t off for cmpext
Despite strange read() API, which mixes these.

+++ src/librados/librados.cc
@@ -140,7 +140,7 @@ void librados::ObjectOperation::set_op_flags2(int flags)
   ::set_op_flags(o, flags);
 }
 
-void librados::ObjectOperation::cmpext(size_t off, uint64_t len,
+void librados::ObjectOperation::cmpext(uint64_t off, size_t len,
                                        bufferlist &cmp_bl,
                                        bufferlist *pmismatch_bl,
                                        int *prval)
